+++ repository/repository.go
@@ -6,7 +6,7 @@ import (
 )
 
 type Repository interface {
-	UserById(ctx context.Context, id int32) (*model.User, error)
+	UserById(ctx context.Context, id int64) (*model.User, error)
 	Create(ctx context.Context, username string) error
 	GetUsers(ctx context.Context)([]*model.User, error)
 }
